chore: moved assertionConcern to utils folder
