+++ src/useCases/produto/index.ts
@@ -2,7 +2,7 @@ import { IProdutoUseCase } from "./produto.interface";
 import { ProdutoDTO } from "./dto";
 import { Produto, CategoriaEnum } from "@entities/produto";
 import { ProdutoGateway } from "@interfaces/gateways/produtoGateway.interface";
-import { AssertionConcern } from "base/assertionConcern";
+import { AssertionConcern } from "@utils/assertionConcern";
 import { ResourceNotFoundError } from "@utils/errors/resourceNotFoundError";
 
 export class ProdutoUseCase implements IProdutoUseCase {
