+++ src/valueObjects/email.ts
@@ -1,5 +1,5 @@
 import { ValueObject } from "@interfaces/valueObject";
-import { AssertionConcern } from "base/assertionConcern";
+import { AssertionConcern } from "@utils/assertionConcern";
 
 interface EmailProperties {
     address: string;
